refactor: create new layout for request pages + use this in access request page
