+++ src/features/amUI/common/RequestPageLayout/RequestPageLayout.module.css
@@ -1,0 +1,27 @@
+.centerBlock {
+  display: block;
+  align-self: center;
+  margin-top: var(--ds-size-4);
+  max-width: 65rem;
+}
+
+.requestBlock {
+  background-color: var(--ds-color-accent-background-default);
+  padding: var(--ds-size-9) var(--ds-size-15);
+}
+
+.headerBlock {
+  display: flex;
+  flex-direction: column;
+  gap: var(--ds-size-2);
+  padding-top: var(--ds-size-14);
+  margin-bottom: 3px;
+}
+
+@media only screen and (max-width: 769px) {
+  .requestBlock {
+    padding-left: var(--ds-size-6);
+    padding-right: var(--ds-size-6);
+    margin-bottom: 6px;
+  }
+}
