+++ src/features/amUI/requestPage/DraftRequestPage/DraftRequestPage.module.css
@@ -1,23 +1,3 @@
-.centerBlock {
-  display: block;
-  align-self: center;
-  margin-top: var(--ds-size-4);
-  max-width: 65rem;
-}
-
-.requestBlock {
-  background-color: var(--ds-color-accent-background-default);
-  padding: var(--ds-size-9) var(--ds-size-15);
-}
-
-.headerBlock {
-  display: flex;
-  flex-direction: column;
-  gap: var(--ds-size-2);
-  padding-top: var(--ds-size-14);
-  margin-bottom: 3px;
-}
-
 .buttonRow {
   display: flex;
   flex-direction: row;
@@ -34,12 +14,4 @@
 .closeWindowInfo {
   font-weight: var(--ds-font-weight-semibold);
   margin-top: var(--ds-size-6);
-}
-
-@media only screen and (max-width: 769px) {
-  .requestBlock {
-    padding-left: var(--ds-size-6);
-    padding-right: var(--ds-size-6);
-    margin-bottom: 6px;
-  }
 }
